Analisis semantico de las asignaciones
