+++ Compilador.h
@@ -1,0 +1,9 @@
+#include "Compilador/TSimbolos.h"
+#include "Compilador/Pila.h"
+#include "Compilador/AnalizadorLexico.h"
+#include "Compilador/AnalizadorSintactico.h"
+#include "Compilador/AnalizadorSemantico.h"
+#include "Compilador/EscribeEnsamblador.h"
+#include "Compilador/AgregaTablaSimbolos.h"
+#include "Compilador/ImprimeTabla.h"
+#include "Compilador/LeeArchivo.h"
