+++ Compilador/TSimbolos.h
