+++ archivofuente.txt
@@ -4,5 +4,5 @@ b=5;
 
 while(a <= b)
 {
-    a = a + 1;
+    a = a / 1;
 }
